+++ tests/net/request.test.ts
@@ -1,24 +1,15 @@
-import { generateKeyPair, ByteBuffer, TxVariant, BodyType, MintTxV0, Request, Asset } from '../../src';
+import { generateKeyPair, ByteBuffer, TxVariant, BodyType, MintTxV0, Asset, Msg, RpcType } from '../../src';
 import Long from 'long';
 
-test('fail to deserialize error request', (): void => {
+test('fail to deserialize invalid request type id', (): void => {
   const buf = ByteBuffer.alloc(128)
     .writeUint32(0)
-    .writeUint8(BodyType.Error)
-    .resetOffset();
-  expect((): void => {
-    Request.deserialize(buf);
-  }).toThrowError('cannot deserialize error requests');
-});
-
-test('fail to deserialize invalid type id request', (): void => {
-  const buf = ByteBuffer.alloc(128)
-    .writeUint32(0)
+    .writeUint8(BodyType.Request)
     .writeUint8(0xff)
     .resetOffset();
   expect((): void => {
-    Request.deserialize(buf);
-  }).toThrowError('unknown type id: 255');
+    Msg.deserialize(buf);
+  }).toThrowError('unknown request id: 255');
 });
 
 test('serialize broadcast request', (): void => {
@@ -43,14 +34,17 @@ test('serialize broadcast request', (): void => {
   tx.sign(generateKeyPair());
   tx.sign(generateKeyPair());
 
-  const req = new Request(123, {
-    type: BodyType.Broadcast,
-    tx,
+  const req = new Msg(123, {
+    type: BodyType.Request,
+    req: {
+      type: RpcType.Broadcast,
+      tx,
+    },
   });
   const buf = ByteBuffer.alloc(4096);
   req.serialize(buf);
   buf.resetOffset();
-  expect(Request.deserialize(buf)).toStrictEqual(req);
+  expect(Msg.deserialize(buf)).toStrictEqual(req);
 });
 
 test('serialize set block filter request', (): void => {
@@ -60,120 +54,150 @@ test('serialize set block filter request', (): void => {
     addrs.push(key.publicKey.toScript().hash());
   }
 
-  const req = new Request(0, {
-    type: BodyType.SetBlockFilter,
-    addrs,
+  const req = new Msg(123, {
+    type: BodyType.Request,
+    req: {
+      type: RpcType.SetBlockFilter,
+      addrs,
+    },
   });
 
   const buf = ByteBuffer.alloc(4096);
   req.serialize(buf);
   buf.resetOffset();
-  expect(Request.deserialize(buf)).toStrictEqual(req);
+  expect(Msg.deserialize(buf)).toStrictEqual(req);
 });
 
 test('serialize set block filter request with no addresses', (): void => {
-  const req = new Request(0, {
-    type: BodyType.SetBlockFilter,
-    addrs: undefined,
+  const req = new Msg(0, {
+    type: BodyType.Request,
+    req: {
+      type: RpcType.SetBlockFilter,
+      addrs: undefined,
+    },
   });
 
   const buf = ByteBuffer.alloc(4096);
   req.serialize(buf);
   buf.resetOffset();
-  expect(Request.deserialize(buf)).toStrictEqual(req);
+  expect(Msg.deserialize(buf)).toStrictEqual(req);
 });
 
 test('serialize clear block filter request', (): void => {
-  const req = new Request(0, {
-    type: BodyType.ClearBlockFilter,
+  const req = new Msg(0, {
+    type: BodyType.Request,
+    req: {
+      type: RpcType.ClearBlockFilter,
+    },
   });
 
   const buf = ByteBuffer.alloc(4096);
   req.serialize(buf);
   buf.resetOffset();
-  expect(Request.deserialize(buf)).toStrictEqual(req);
+  expect(Msg.deserialize(buf)).toStrictEqual(req);
 });
 
 test('serialize subscribe request', (): void => {
-  const req = new Request(0, {
-    type: BodyType.Subscribe,
+  const req = new Msg(0, {
+    type: BodyType.Request,
+    req: {
+      type: RpcType.Subscribe,
+    },
   });
 
   const buf = ByteBuffer.alloc(4096);
   req.serialize(buf);
   buf.resetOffset();
-  expect(Request.deserialize(buf)).toStrictEqual(req);
+  expect(Msg.deserialize(buf)).toStrictEqual(req);
 });
 
 test('serialize unsubscribe request', (): void => {
-  const req = new Request(0, {
-    type: BodyType.Unsubscribe,
+  const req = new Msg(0, {
+    type: BodyType.Request,
+    req: {
+      type: RpcType.Unsubscribe,
+    },
   });
 
   const buf = ByteBuffer.alloc(4096);
   req.serialize(buf);
   buf.resetOffset();
-  expect(Request.deserialize(buf)).toStrictEqual(req);
+  expect(Msg.deserialize(buf)).toStrictEqual(req);
 });
 
 test('serialize get properties request', (): void => {
-  const req = new Request(0, {
-    type: BodyType.GetProperties,
+  const req = new Msg(0, {
+    type: BodyType.Request,
+    req: {
+      type: RpcType.GetProperties,
+    },
   });
 
   const buf = ByteBuffer.alloc(4096);
   req.serialize(buf);
   buf.resetOffset();
-  expect(Request.deserialize(buf)).toStrictEqual(req);
+  expect(Msg.deserialize(buf)).toStrictEqual(req);
 });
 
 test('serialize get block request', (): void => {
-  const req = new Request(0, {
-    type: BodyType.GetBlock,
-    height: Long.fromNumber(123, true),
+  const req = new Msg(0, {
+    type: BodyType.Request,
+    req: {
+      type: RpcType.GetBlock,
+      height: Long.fromNumber(123, true),
+    },
   });
 
   const buf = ByteBuffer.alloc(4096);
   req.serialize(buf);
   buf.resetOffset();
-  expect(Request.deserialize(buf)).toStrictEqual(req);
+  expect(Msg.deserialize(buf)).toStrictEqual(req);
 });
 
 test('serialize get full block request', (): void => {
-  const req = new Request(0, {
-    type: BodyType.GetFullBlock,
-    height: Long.fromNumber(123, true),
+  const req = new Msg(0, {
+    type: BodyType.Request,
+    req: {
+      type: RpcType.GetFullBlock,
+      height: Long.fromNumber(123, true),
+    },
   });
 
   const buf = ByteBuffer.alloc(4096);
   req.serialize(buf);
   buf.resetOffset();
-  expect(Request.deserialize(buf)).toStrictEqual(req);
+  expect(Msg.deserialize(buf)).toStrictEqual(req);
 });
 
 test('serialize get block range request', (): void => {
-  const req = new Request(0, {
-    type: BodyType.GetBlockRange,
-    minHeight: Long.fromNumber(123, true),
-    maxHeight: Long.fromNumber(456, true),
+  const req = new Msg(0, {
+    type: BodyType.Request,
+    req: {
+      type: RpcType.GetBlockRange,
+      minHeight: Long.fromNumber(123, true),
+      maxHeight: Long.fromNumber(456, true),
+    },
   });
 
   const buf = ByteBuffer.alloc(4096);
   req.serialize(buf);
   buf.resetOffset();
-  expect(Request.deserialize(buf)).toStrictEqual(req);
+  expect(Msg.deserialize(buf)).toStrictEqual(req);
 });
 
 test('serialize get address info request', (): void => {
-  const req = new Request(0, {
-    type: BodyType.GetAddressInfo,
-    addr: generateKeyPair()
-      .publicKey.toScript()
-      .hash(),
+  const req = new Msg(0, {
+    type: BodyType.Request,
+    req: {
+      type: RpcType.GetAddressInfo,
+      addr: generateKeyPair()
+        .publicKey.toScript()
+        .hash(),
+    },
   });
 
   const buf = ByteBuffer.alloc(4096);
   req.serialize(buf);
   buf.resetOffset();
-  expect(Request.deserialize(buf)).toStrictEqual(req);
+  expect(Msg.deserialize(buf)).toStrictEqual(req);
 });